instrumenting inverse wisharts.
the naivesampler and the snappysampler don't seem to quiiite agree in some marginals. We'll have to look into this.

Current timings out of NIW.R:
on ~1e5 samples: (forgive the accidentally negatives)
Loading required package: MASS
   user  system elapsed
-12.704   0.000 -12.721
   user  system elapsed
 -5.270  -0.004  -5.279

on ~1e6 samples:
   user  system elapsed
 127.16    0.03  127.48
   user  system elapsed
 55.200   0.037  55.339

My prediction in mathematics.lyx that the improved algorithm is
only a piddly 2-fold speedup is panning out nicely...
There's still some optimizations to be made, though.