omg i fixed it
it turned out the bug with the bad plots was because, what else, apply() is confusing and difficult
I switched to operating on the vectors one by one, which I already had.
That's what I get for trying to be clever.
This probably isn't even any slower, except that maybe mean() could be done in C.