bugthree: chol() the appropriate place.
at least for j=0, this gives a distribution matching multinormal(),
which basically means I've gotten all my syntax in the right order,
but not that the gibbs part is working, since in the j=0 case the
gibbs sampler collapses to just a really slow standard multinormal sampler.

the new guards are since chol() is less tolerant of empty vectors than dot().

I still don't understand how the *derivation* of this came about
 and therefore i can't debug it any further :(